// src/app/model/Aluno.ts
export class Aluno {
  alunov: number;
  ano: string;
  anocon: string;
  autate: number;
  baiempres: string;
  baimae: string;
  baipai: string;
  bairespon: string;
  bairro: string;
  banco_PADRAO: number;
  bloquear_MATRICULA: number;
  bloquear_PORTAL: number;
  celrespon: string;
  cep: string;
  cepmae: string;
  ceppai: string;
  ceprespon: string;
  cesao: string;
  cid: string;
  cidmae: string;
  cidpai: string;
  cidrespon: string;
  comple: string;
  complemae: string;
  complepai: string;
  compleresp: string;
  contat: string;
  cor: string;
  cpfmae: string;
  cpfpai: string;
  cpfrespon: string;
  ctbel: number;
  curso_pos: string;
  datcad: number;
  datcon: number;
  datnas: number;
  datnasmae: string;
  datnaspai: string;
  datnasresp: number;
  datsitbib: number;
  ddd: string;
  deficiencia: number;
  diappg: number;
  email: string;
  emailmae: string;
  emailpai: string;
  emailrespon: string;
  empmae: string;
  emppai: string;
  emprespon: string;
  endempres: string;
  endereco: string;
  endmae: string;
  endpai: string;
  endrespon: string;
  escola: number;
  estciv: string;
  estcivrespon: string;
  foncomrespon: string;
  fone: string;
  fonecel: string;
  fonmae: string;
  fonpai: string;
  fonrespon: string;
  forma_INGRESSO: string;
  forori: string;
  id: number;
  internmae: number;
  internpai: number;
  locnas: string;
  mae: string;
  maerespon: string;
  matric: number;
  mensagem_BIBLIOTECA: string;
  mensagem_PORTAL: string;
  nacion: string;
  nacionrespon: string;
  natura: string;
  nome: string;
  nome_censo: string;
  nregoco: number;
  numcpf: string;
  numero: string;
  numero_RESPON: string;
  numrg: string;
  obs: string;
  obssitbib: string;
  opeaut: string;
  opesitbib: string;
  pai: string;
  pairespon: string;
  pendencia_CENSO: string;
  promae: string;
  propai: string;
  prorespon: string;
  religi: string;
  respon: string;
  rgmae: string;
  rgorgao: string;
  rgorgaorespon: string;
  rgpai: string;
  rgrespon: string;
  senha: string;
  seqano: string;
  serasa_DATA: number;
  serasa_ID_OBS: string;
  serasa_OPERADOR: string;
  serasa_OPERADOR_DIA: number;
  serasa_REMOVIDO: number;
  serasa_REMOVIDO_OPERADOR: string;
  sexo: string;
  sitbib: string;
  status: string;
  titele: string;
  trocacurso: string;
  uf: string;
  ufmae: string;
  ufpai: string;
  ufrespon: string;
  ultaut: number;
  ultbai: string;
  ultcargo: string;
  ultcep: string;
  ultcid: string;
  ultcomple: string;
  ultcontat: string;
  ultend: string;
  ultest: string;
  ultfon: string;
  ultgra: string;
  ultser: string;
  ultuf: string;
  zona: string
}